+++ maiacore/include/maiacore/config.h
@@ -1,0 +1,10 @@
+enum class TuningSystem {
+    EQUAL_TEMPERAMENT,
+    JUST_INTONATION,
+    PYTHAGOREAN_TUNING,
+    MEANTONE_TEMPERAMENT,
+    WELL_TEMPERAMENT
+};
+
+TuningSystem getTuningSystem();
+void setTuningSystem(TuningSystem tuningSystem);