+++ maiacore/src/maiacore/config.cpp
@@ -1,0 +1,6 @@
+#include "maiacore/config.h"
+
+TuningSystem _tuningSystem = TuningSystem::EQUAL_TEMPERAMENT;
+
+TuningSystem getTuningSystem() { return _tuningSystem; }
+void setTuningSystem(TuningSystem tuningSystem) { _tuningSystem = tuningSystem; }