+++ maiacore/src/maiacore/python_wrapper/py_interval.cpp
@@ -137,6 +137,8 @@ void IntervalClass(const py::module& m) {
     cls.def("isAnyOctaveSeventh", &Interval::isAnyOctaveSeventh);
     cls.def("isAnyOctaveOctave", &Interval::isAnyOctaveOctave);
 
+    cls.def("toCents", &Interval::toCents);
+
     // Default Python 'print' function:
     cls.def("__repr__", [](const Interval& interval) {
         const std::string direction = (interval.isAscendant()) ? "asc" : "desc";