+++ maialib/maiapy/__init__.py
@@ -1,3 +1,3 @@
 from .other import *
 from .plots import *
-from .plotPartsActivity import plotPartsActivity
+from .plotPartsActivity import *